functions/train-model/requirements.txt
absl-py==2.0.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
argcomplete==1.10.3
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.8.2
cachetools==5.3.2
certifi==2023.11.17
chardet==3.0.4
charset-normalizer==3.3.2
compressed-rtf==1.0.6
dataclasses-json==0.6.3
distro==1.9.0
docx2txt==0.8
ebcdic==1.1.1
extract-msg==0.28.7
faiss-cpu==1.7.4
filelock==3.13.1
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
h5py==3.10.0
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.1
idna==3.6
IMAPClient==2.1.0
jsonpatch==1.33
jsonpointer==2.4
keras==2.15.0
langchain==0.0.353
langchain-community==0.0.7
langchain-core==0.1.4
langsmith==0.0.75
libclang==16.0.6
lxml==5.0.0
Markdown==3.5.1
MarkupSafe==2.1.3
marshmallow==3.20.1
ml-dtypes==0.2.0
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.2
oauthlib==3.2.2
olefile==0.47
openai==1.6.1
opt-einsum==3.3.0
packaging==23.2
pdfminer.six==20191110
pillow==10.2.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodome==3.19.1
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-pptx==0.6.23
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.1
six==1.12.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.24
tenacity==8.2.3
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
termcolor==2.4.0
textract==1.6.5
tiktoken==0.5.2
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.36.2
typing-inspect==0.9.0
typing_extensions==4.9.0
tzlocal==5.2
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
xlrd==1.2.0
XlsxWriter==3.1.9
yarl==1.9.4
